+++ setuptools_rust/__init__.py
@@ -1,6 +1,6 @@
 from .build import build_rust
 from .clean import clean_rust
-from .extension import Binding, RustExtension, Strip
+from .extension import Binding, RustBin, RustExtension, Strip
 from .version import version as __version__
 
-__all__ = ("Binding", "RustExtension", "Strip", "build_rust", "clean_rust")
+__all__ = ("Binding", "RustBin", "RustExtension", "Strip", "build_rust", "clean_rust")